backend/requirements.txt
sanic
sanic-ext
tortoise-orm
aiomysql
cryptography
redis
bcrypt
pyjwt
python-dotenv
requests
pytest
aerich
sanic-limiter
aiohttp
aiosmtplib
selenium
webdriver-manager
